feat: admin crud allowed permission

+++ todolist/app/Http/Controllers/TodoController.php
@@ -32,6 +32,10 @@ class TodoController extends Controller
             return response()->json(['error' => 'User not found'], 404);
         }
 
+        if (!$user->is_crud_allowed) {
+            return response()->json(['error' => 'Unauthorized'], 401);
+        }
+
         $todo = new Todo($request->all());
         $todo->user_id = $userId; 
         $user->todos()->save($todo);
@@ -45,6 +49,11 @@ class TodoController extends Controller
             return response()->json(['error' => 'Unauthorized'], 401);
         }
 
+        $user = User::find($userId);
+        if (!$user || !$user->is_crud_allowed) {
+            return response()->json(['error' => 'Unauthorized'], 401);
+        }
+
         $todo->update($request->all());
         return $todo;
     }
@@ -55,7 +64,13 @@ class TodoController extends Controller
             return response()->json(['error' => 'Unauthorized'], 401);
         }
 
+        $user = User::find($userId);
+        if (!$user || !$user->is_crud_allowed) {
+            return response()->json(['error' => 'Unauthorized'], 401);
+        }
+
         $todo->delete();
         return response()->json(null, 204);
     }
 }
+
